requirements.txt
torch>=2.1
numpy>=1.23
tqdm>=4.66
Pillow>=10.0
imageio>=2.31
imageio-ffmpeg>=0.4.8
